services/storage-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
boto3==1.34.10
